requirements.txt
Django==1.9.4
Pillow==3.1.1
PyJWT==1.4.0
Unidecode==0.4.19
argparse==1.2.1
dj-database-url==0.4.1
dj-static==0.0.6
django-braces==1.8.1
django-crispy-forms==1.6.0
django-filter==0.13.0
django-model-utils==2.4
django-registration-redux==1.4
django-toolbelt==0.0.1
gunicorn==19.4.5
httplib2==0.9.2
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
static3==0.6.1
whitenoise==3.0
wsgiref==0.1.2
